refactor(irb)!: Phase 4 slice 7d — the IRB namespace becomes plain typed functions; the 6 disabled ty rules are back on
engine/irb/namespace.py (984 LOC, IRBLazyFrame + IRBExpr) deleted;
engine/irb/transforms.py lands all 20 transforms with bodies verbatim
(scaling-factor reconstructions deliberately NOT rewritten — Phase 5
rulepack threading). ~369 call sites converted (7 src + ~362 test
accessors across 24 files; test_irb_namespace.py renamed to
test_irb_transforms.py). Zero residual namespace references;
rg register_*_namespace over src/ returns nothing — the Polars namespace
pattern is extinct from first-party code.

With the last registration gone, the [tool.ty.rules] disabling block is
deleted: unresolved-attribute, invalid-argument-type, invalid-assignment,
invalid-parameter-default, no-matching-overload and not-subscriptable are
enforced project-wide again. Burn-down 1,366 -> 0 diagnostics, typing-only:
ColumnSpec/EdgeColumn dtype widened to PolarsDataType (~1,050), when-then
chain aliases, has_required_columns -> TypeGuard, brand made generic over
frame type, COREP/Pillar-3 workbook typing, ty lock 0.0.26 -> 0.0.49
(resolves polars collect() overloads). 9 per-line ignores, all the same
justified class (config.irb_permissions derived non-None in __post_init__).
max_engine_module_loc 1,494 -> 1,499 banked (+5 lines of typing
infrastructure, justified).

Slice 7 complete: ccr -> slotting -> sa -> irb all retired, each
parity-byte-identical. Gates: unit+contracts 6,066; acceptance+integration
1,482; parity byte-identical; ruff/ty (all rules)/arch_check green.
