+++ scripts/arch_check.py
@@ -152,7 +152,6 @@ LOGGER_REQUIRED_EXEMPT: set[str] = {
     "engine/irb/adjustments.py",
     "engine/irb/formulas.py",
     "engine/irb/guarantee.py",
-    "engine/irb/namespace.py",
     "engine/irb/stats_backend.py",
     "engine/aggregator/_crm_reporting.py",
     "engine/aggregator/_el_summary.py",
