+++ scripts/arch_metrics.json
@@ -4,6 +4,6 @@
   "engine_presence_guard_sites": 372,
   "engine_collect_schema_sites": 166,
   "engine_eager_collect_sites": 38,
-  "max_engine_module_loc": 1494,
+  "max_engine_module_loc": 1499,
   "cites_decorators": 193
 }
