+++ src/rwa_calc/engine/irb/__init__.py
@@ -3,22 +3,27 @@
 Provides:
 - IRBCalculator: Main IRB calculator implementing IRBCalculatorProtocol
 - IRB formulas: K formula, correlation, maturity adjustment
-- IRBLazyFrame: Polars namespace for fluent IRB calculations
-- IRBExpr: Polars expression namespace for column-level operations
+- Plain typed IRB transforms in the sibling ``transforms`` module,
+  composed via ``LazyFrame.pipe``
 
 Supports both F-IRB (supervisory LGD) and A-IRB (own LGD estimates).
 
-Usage with namespace:
+Usage with transforms:
     import polars as pl
     from rwa_calc.contracts.config import CalculationConfig
-    from rwa_calc.engine.irb import IRBLazyFrame  # Registers namespace
+    from rwa_calc.engine.irb.transforms import (
+        apply_all_formulas,
+        apply_firb_lgd,
+        classify_approach,
+        prepare_columns,
+    )
 
     config = CalculationConfig.crr(reporting_date=date(2024, 12, 31))
-    result = (lf
-        .irb.classify_approach(config)
-        .irb.apply_firb_lgd(config)
-        .irb.prepare_columns(config)
-        .irb.apply_all_formulas(config)
+    result = (
+        lf.pipe(classify_approach, config)
+        .pipe(apply_firb_lgd, config)
+        .pipe(prepare_columns, config)
+        .pipe(apply_all_formulas, config)
     )
 
 References:
@@ -27,8 +32,6 @@ References:
 - CRR Art. 162-163: Maturity and PD floors
 """
 
-# Import namespace module to register namespaces on module load
-import rwa_calc.engine.irb.namespace  # noqa: F401
 from rwa_calc.engine.irb.calculator import IRBCalculator, create_irb_calculator
 from rwa_calc.engine.irb.formulas import (
     apply_irb_formulas,
@@ -38,7 +41,6 @@ from rwa_calc.engine.irb.formulas import (
     calculate_k,
     calculate_maturity_adjustment,
 )
-from rwa_calc.engine.irb.namespace import IRBExpr, IRBLazyFrame
 
 __all__ = [
     "IRBCalculator",
@@ -49,7 +51,4 @@ __all__ = [
     "calculate_irb_rwa",
     "calculate_expected_loss",
     "apply_irb_formulas",
-    # Namespace classes
-    "IRBLazyFrame",
-    "IRBExpr",
 ]
