+++ tests/acceptance/basel31/test_scenario_b31_d7_parameter_substitution.py
@@ -37,11 +37,18 @@ from datetime import date
 import polars as pl
 import pytest
 
-import rwa_calc.engine.irb.namespace  # noqa: F401 - Register namespace
 from rwa_calc.contracts.config import CalculationConfig
 from rwa_calc.domain.enums import ApproachType, PermissionMode
 from rwa_calc.engine.crm.processor import CRMProcessor
 from rwa_calc.engine.irb.formulas import _parametric_irb_risk_weight_expr
+from rwa_calc.engine.irb.transforms import (
+    apply_all_formulas,
+    apply_firb_lgd,
+    apply_guarantee_substitution,
+    classify_approach,
+    compute_el_shortfall_excess,
+    prepare_columns,
+)
 from tests.fixtures.contract_columns import pad_crm_exit_defaults as _pad
 
 
@@ -167,12 +174,12 @@ def _create_crm_and_irb_result(
     # Run full IRB formula pipeline and guarantee substitution
     result = (
         _pad(exposures_with_guarantee)
-        .irb.classify_approach(config)
-        .irb.apply_firb_lgd(config)
-        .irb.prepare_columns(config)
-        .irb.apply_all_formulas(config)
-        .irb.compute_el_shortfall_excess()
-        .irb.apply_guarantee_substitution(config)
+        .pipe(classify_approach, config)
+        .pipe(apply_firb_lgd, config)
+        .pipe(prepare_columns, config)
+        .pipe(apply_all_formulas, config)
+        .pipe(compute_el_shortfall_excess)
+        .pipe(apply_guarantee_substitution, config)
     )
 
     return result.collect()
