+++ tests/acceptance/crr/test_scenario_crr_i_defaulted.py
@@ -27,7 +27,10 @@ from tests.fixtures.contract_columns import pad_crm_exit_defaults as _pad
 
 from rwa_calc.contracts.config import CalculationConfig
 from rwa_calc.domain.enums import PermissionMode
-from rwa_calc.engine.irb import IRBLazyFrame  # noqa: F401 - registers namespace
+from rwa_calc.engine.irb.transforms import (
+    apply_all_formulas,
+    prepare_columns,
+)
 
 # =============================================================================
 # Fixtures
@@ -97,7 +100,9 @@ class TestCRRI1_FIRBCorporateDefaulted:
             ead_final=500_000.0,
         )
         result = (
-            lf.irb.prepare_columns(crr_irb_config).irb.apply_all_formulas(crr_irb_config).collect()
+            lf.pipe(prepare_columns, crr_irb_config)
+            .pipe(apply_all_formulas, crr_irb_config)
+            .collect()
         )
         assert result["k"][0] == pytest.approx(0.0, abs=1e-10)
 
@@ -113,7 +118,9 @@ class TestCRRI1_FIRBCorporateDefaulted:
             ead_final=500_000.0,
         )
         result = (
-            lf.irb.prepare_columns(crr_irb_config).irb.apply_all_formulas(crr_irb_config).collect()
+            lf.pipe(prepare_columns, crr_irb_config)
+            .pipe(apply_all_formulas, crr_irb_config)
+            .collect()
         )
         assert result["rwa"][0] == pytest.approx(0.0, abs=1e-6)
         assert result["risk_weight"][0] == pytest.approx(0.0, abs=1e-6)
@@ -130,7 +137,9 @@ class TestCRRI1_FIRBCorporateDefaulted:
             ead_final=500_000.0,
         )
         result = (
-            lf.irb.prepare_columns(crr_irb_config).irb.apply_all_formulas(crr_irb_config).collect()
+            lf.pipe(prepare_columns, crr_irb_config)
+            .pipe(apply_all_formulas, crr_irb_config)
+            .collect()
         )
         assert result["expected_loss"][0] == pytest.approx(225_000.0, rel=1e-6)
 
@@ -161,7 +170,9 @@ class TestCRRI2_AIRBRetailDefaulted:
             ead_final=25_000.0,
         )
         result = (
-            lf.irb.prepare_columns(crr_irb_config).irb.apply_all_formulas(crr_irb_config).collect()
+            lf.pipe(prepare_columns, crr_irb_config)
+            .pipe(apply_all_formulas, crr_irb_config)
+            .collect()
         )
         assert result["k"][0] == pytest.approx(0.15, abs=1e-10)
 
@@ -177,7 +188,9 @@ class TestCRRI2_AIRBRetailDefaulted:
             ead_final=25_000.0,
         )
         result = (
-            lf.irb.prepare_columns(crr_irb_config).irb.apply_all_formulas(crr_irb_config).collect()
+            lf.pipe(prepare_columns, crr_irb_config)
+            .pipe(apply_all_formulas, crr_irb_config)
+            .collect()
         )
         expected_rwa = 0.15 * 12.5 * 1.0 * 25_000.0  # 46,875
         assert result["rwa"][0] == pytest.approx(expected_rwa, rel=1e-6)
@@ -194,7 +207,9 @@ class TestCRRI2_AIRBRetailDefaulted:
             ead_final=25_000.0,
         )
         result = (
-            lf.irb.prepare_columns(crr_irb_config).irb.apply_all_formulas(crr_irb_config).collect()
+            lf.pipe(prepare_columns, crr_irb_config)
+            .pipe(apply_all_formulas, crr_irb_config)
+            .collect()
         )
         assert result["expected_loss"][0] == pytest.approx(12_500.0, rel=1e-6)
 
@@ -229,7 +244,9 @@ class TestCRRI3_AIRBCorporateDefaulted:
             ead_final=500_000.0,
         )
         result = (
-            lf.irb.prepare_columns(crr_irb_config).irb.apply_all_formulas(crr_irb_config).collect()
+            lf.pipe(prepare_columns, crr_irb_config)
+            .pipe(apply_all_formulas, crr_irb_config)
+            .collect()
         )
         assert result["k"][0] == pytest.approx(0.15, abs=1e-10)
 
@@ -248,7 +265,9 @@ class TestCRRI3_AIRBCorporateDefaulted:
             ead_final=500_000.0,
         )
         result = (
-            lf.irb.prepare_columns(crr_irb_config).irb.apply_all_formulas(crr_irb_config).collect()
+            lf.pipe(prepare_columns, crr_irb_config)
+            .pipe(apply_all_formulas, crr_irb_config)
+            .collect()
         )
         expected_rwa = 0.15 * 12.5 * 500_000.0  # 937,500
         assert result["rwa"][0] == pytest.approx(expected_rwa, rel=1e-6)
@@ -265,6 +284,8 @@ class TestCRRI3_AIRBCorporateDefaulted:
             ead_final=500_000.0,
         )
         result = (
-            lf.irb.prepare_columns(crr_irb_config).irb.apply_all_formulas(crr_irb_config).collect()
+            lf.pipe(prepare_columns, crr_irb_config)
+            .pipe(apply_all_formulas, crr_irb_config)
+            .collect()
         )
         assert result["expected_loss"][0] == pytest.approx(225_000.0, rel=1e-6)
