+++ tests/unit/test_b31_revolving_maturity.py
@@ -20,7 +20,10 @@ import polars as pl
 import pytest
 
 from rwa_calc.contracts.config import CalculationConfig
-from rwa_calc.engine.irb import IRBExpr, IRBLazyFrame  # noqa: F401 - registers namespace
+from rwa_calc.engine.irb.transforms import (
+    apply_all_formulas,
+    prepare_columns,
+)
 from tests.fixtures.contract_columns import pad_crm_exit_defaults as _pad
 
 # =============================================================================
@@ -67,7 +70,7 @@ class TestB31RevolvingMaturity:
             }
         )
 
-        result = _pad(lf).irb.prepare_columns(b31_config).collect()
+        result = _pad(lf).pipe(prepare_columns, b31_config).collect()
         # M should be 5.0 (capped), derived from facility_termination_date
         assert result["maturity"][0] == pytest.approx(5.0, abs=0.05)
 
@@ -85,7 +88,7 @@ class TestB31RevolvingMaturity:
             }
         )
 
-        result = _pad(lf).irb.prepare_columns(b31_config).collect()
+        result = _pad(lf).pipe(prepare_columns, b31_config).collect()
         # M should be ~1.0, derived from maturity_date (not termination date)
         assert result["maturity"][0] == pytest.approx(1.0, abs=0.05)
 
@@ -104,7 +107,7 @@ class TestB31RevolvingMaturity:
             }
         )
 
-        result = _pad(lf).irb.prepare_columns(crr_config).collect()
+        result = _pad(lf).pipe(prepare_columns, crr_config).collect()
         # CRR ignores termination date — M from maturity_date ≈ 1.0
         assert result["maturity"][0] == pytest.approx(1.0, abs=0.05)
 
@@ -124,7 +127,7 @@ class TestB31RevolvingMaturity:
             }
         ).cast({"facility_termination_date": pl.Date})
 
-        result = _pad(lf).irb.prepare_columns(b31_config).collect()
+        result = _pad(lf).pipe(prepare_columns, b31_config).collect()
         # Null termination date → fall back to maturity_date → M ≈ 3.0
         assert result["maturity"][0] == pytest.approx(3.0, abs=0.05)
 
@@ -143,7 +146,7 @@ class TestB31RevolvingMaturity:
             }
         )
 
-        result = _pad(lf).irb.prepare_columns(b31_config).collect()
+        result = _pad(lf).pipe(prepare_columns, b31_config).collect()
         # No termination date column → use maturity_date → M ≈ 3.0
         assert result["maturity"][0] == pytest.approx(3.0, abs=0.05)
 
@@ -161,7 +164,7 @@ class TestB31RevolvingMaturity:
             }
         )
 
-        result = _pad(lf).irb.prepare_columns(b31_config).collect()
+        result = _pad(lf).pipe(prepare_columns, b31_config).collect()
         # 10 years from reporting but capped at 5.0
         assert result["maturity"][0] == pytest.approx(5.0, abs=0.01)
 
@@ -179,7 +182,7 @@ class TestB31RevolvingMaturity:
             }
         )
 
-        result = _pad(lf).irb.prepare_columns(b31_config).collect()
+        result = _pad(lf).pipe(prepare_columns, b31_config).collect()
         # 0.25yr from reporting but floored at 1.0
         assert result["maturity"][0] == pytest.approx(1.0, abs=0.01)
 
@@ -205,7 +208,7 @@ class TestB31RevolvingMaturity:
             }
         ).cast({"facility_termination_date": pl.Date})
 
-        result = _pad(lf).irb.prepare_columns(b31_config).collect()
+        result = _pad(lf).pipe(prepare_columns, b31_config).collect()
         # Row 0: revolving + termination date → M ≈ 4.0
         assert result["maturity"][0] == pytest.approx(4.0, abs=0.05)
         # Row 1: non-revolving → M ≈ 1.0 (from maturity_date)
@@ -229,7 +232,7 @@ class TestB31RevolvingMaturity:
             }
         ).cast({"is_revolving": pl.Boolean})
 
-        result = _pad(lf).irb.prepare_columns(b31_config).collect()
+        result = _pad(lf).pipe(prepare_columns, b31_config).collect()
         # Null is_revolving → treated as non-revolving → M from maturity_date ≈ 1.0
         assert result["maturity"][0] == pytest.approx(1.0, abs=0.05)
 
@@ -255,8 +258,8 @@ class TestB31RevolvingMaturity:
         )
         rwa_non_revolving = (
             _pad(lf_non_revolving)
-            .irb.prepare_columns(b31_config)
-            .irb.apply_all_formulas(b31_config)
+            .pipe(prepare_columns, b31_config)
+            .pipe(apply_all_formulas, b31_config)
             .collect()["rwa"][0]
         )
 
@@ -271,8 +274,8 @@ class TestB31RevolvingMaturity:
         )
         rwa_revolving = (
             _pad(lf_revolving)
-            .irb.prepare_columns(b31_config)
-            .irb.apply_all_formulas(b31_config)
+            .pipe(prepare_columns, b31_config)
+            .pipe(apply_all_formulas, b31_config)
             .collect()["rwa"][0]
         )
 
@@ -296,7 +299,10 @@ class TestB31RevolvingMaturity:
         )
 
         result = (
-            _pad(lf).irb.prepare_columns(b31_config).irb.apply_all_formulas(b31_config).collect()
+            _pad(lf)
+            .pipe(prepare_columns, b31_config)
+            .pipe(apply_all_formulas, b31_config)
+            .collect()
         )
         # Retail always gets MA = 1.0 regardless of maturity source
         assert result["maturity_adjustment"][0] == pytest.approx(1.0)
@@ -316,6 +322,6 @@ class TestB31RevolvingMaturity:
             }
         )
 
-        result = _pad(lf).irb.prepare_columns(b31_config).collect()
+        result = _pad(lf).pipe(prepare_columns, b31_config).collect()
         # No is_revolving column → treated as non-revolving → M from maturity_date ≈ 1.0
         assert result["maturity"][0] == pytest.approx(1.0, abs=0.05)
