+++ tests/unit/test_basel31_engine.py
@@ -26,7 +26,6 @@ from decimal import Decimal
 import polars as pl
 import pytest
 
-import rwa_calc.engine.irb.namespace  # noqa: F401 - register namespace
 from rwa_calc.contracts.config import CalculationConfig
 from rwa_calc.data.tables.firb_lgd import (
     BASEL31_FIRB_SUPERVISORY_LGD,
@@ -45,6 +44,13 @@ from rwa_calc.engine.irb.formulas import (
     calculate_correlation,
 )
 from rwa_calc.engine.irb.formulas import apply_irb_formulas as _apply_irb_formulas_raw
+from rwa_calc.engine.irb.transforms import (
+    apply_all_formulas,
+    apply_firb_lgd,
+    apply_lgd_floor,
+    apply_pd_floor,
+    prepare_columns,
+)
 from tests.fixtures.contract_columns import pad_crm_exit_defaults as _pad
 from tests.fixtures.single_exposure import calculate_single_equity_exposure
 
@@ -333,7 +339,7 @@ class TestPDFloors:
                 "pd": [0.0001],
             }
         )
-        result = _pad(lf).irb.apply_pd_floor(crr_config).collect()
+        result = _pad(lf).pipe(apply_pd_floor, crr_config).collect()
         assert result["pd_floored"][0] == pytest.approx(0.0003)
 
     def test_namespace_apply_pd_floor_basel31(
@@ -347,7 +353,7 @@ class TestPDFloors:
                 "pd": [0.0001],
             }
         )
-        result = _pad(lf).irb.apply_pd_floor(basel31_config).collect()
+        result = _pad(lf).pipe(apply_pd_floor, basel31_config).collect()
         # QRRE revolver default is 0.10%
         assert result["pd_floored"][0] == pytest.approx(0.0010)
 
@@ -366,8 +372,8 @@ class TestPDFloors:
         )
         result = (
             _pad(lf)
-            .irb.prepare_columns(basel31_config)
-            .irb.apply_all_formulas(basel31_config)
+            .pipe(prepare_columns, basel31_config)
+            .pipe(apply_all_formulas, basel31_config)
             .collect()
         )
         assert result["pd_floored"][0] == pytest.approx(0.0005)
@@ -734,7 +740,7 @@ class TestLGDFloors:
                 "lgd": [0.10],
             }
         )
-        result = _pad(lf).irb.apply_lgd_floor(crr_config).collect()
+        result = _pad(lf).pipe(apply_lgd_floor, crr_config).collect()
         assert result["lgd_floored"][0] == pytest.approx(0.10)
 
     def test_namespace_apply_lgd_floor_basel31_airb(
@@ -750,7 +756,7 @@ class TestLGDFloors:
                 "is_airb": [True],
             }
         )
-        result = _pad(lf).irb.apply_lgd_floor(basel31_config).collect()
+        result = _pad(lf).pipe(apply_lgd_floor, basel31_config).collect()
         assert result["lgd_floored"][0] == pytest.approx(0.25)
 
     def test_namespace_apply_lgd_floor_basel31_firb_no_floor(
@@ -766,7 +772,7 @@ class TestLGDFloors:
                 "is_airb": [False],
             }
         )
-        result = _pad(lf).irb.apply_lgd_floor(basel31_config).collect()
+        result = _pad(lf).pipe(apply_lgd_floor, basel31_config).collect()
         # F-IRB: no floor applied
         assert result["lgd_floored"][0] == pytest.approx(0.10)
 
@@ -984,7 +990,7 @@ class TestLGDFloors:
                 "is_airb": [True, True],
             }
         )
-        result = _pad(lf).irb.apply_lgd_floor(basel31_config).collect()
+        result = _pad(lf).pipe(apply_lgd_floor, basel31_config).collect()
         # Both get 25% floor — corporate subordinated same as senior (Art. 161(5))
         assert result["lgd_floored"][0] == pytest.approx(0.25)
         assert result["lgd_floored"][1] == pytest.approx(0.25)
@@ -1135,7 +1141,7 @@ class TestLGDFloors:
                 "is_airb": [True, True, True],
             }
         )
-        result = _pad(lf).irb.apply_lgd_floor(basel31_config).collect()
+        result = _pad(lf).pipe(apply_lgd_floor, basel31_config).collect()
         # Retail other: 30% (Art. 164(4)(b)(ii))
         assert result["lgd_floored"][0] == pytest.approx(0.30)
         # Retail QRRE: 50% (Art. 164(4)(b)(i))
@@ -1155,7 +1161,7 @@ class TestLGDFloors:
                 "is_airb": [True],
             }
         )
-        result = _pad(lf).irb.apply_lgd_floor(basel31_config).collect()
+        result = _pad(lf).pipe(apply_lgd_floor, basel31_config).collect()
         assert result["lgd_floored"][0] == pytest.approx(0.05)
 
     def test_retail_lgd_above_floor_unchanged(
@@ -1173,7 +1179,7 @@ class TestLGDFloors:
                 "is_airb": [True],
             }
         )
-        result = _pad(lf).irb.apply_lgd_floor(basel31_config).collect()
+        result = _pad(lf).pipe(apply_lgd_floor, basel31_config).collect()
         assert result["lgd_floored"][0] == pytest.approx(0.40)
 
     # --- P1.8: Generic 'immovable' collateral_type routing for RRE/CRE ---
@@ -1343,7 +1349,7 @@ class TestLGDFloors:
                 "collateral_type": ["immovable"],
             }
         )
-        result = _pad(lf).irb.apply_lgd_floor(basel31_config).collect()
+        result = _pad(lf).pipe(apply_lgd_floor, basel31_config).collect()
         assert result["lgd_floored"][0] == pytest.approx(0.05)
 
     def test_mixed_immovable_collateral_types_batch(
@@ -1403,8 +1409,8 @@ class TestLGDFloors:
         )
         result = (
             _pad(lf)
-            .irb.prepare_columns(basel31_config)
-            .irb.apply_all_formulas(basel31_config)
+            .pipe(prepare_columns, basel31_config)
+            .pipe(apply_all_formulas, basel31_config)
             .collect()
         )
         # A-IRB: LGD floored to 25%
@@ -1653,7 +1659,7 @@ class TestFIRBSupervisoryLGD:
                 "approach": [ApproachType.FIRB.value],
             }
         )
-        result = _pad(lf).irb.apply_firb_lgd(crr_config).collect()
+        result = _pad(lf).pipe(apply_firb_lgd, crr_config).collect()
         assert result["lgd"][0] == pytest.approx(0.45)
 
     def test_namespace_apply_firb_lgd_basel31(
@@ -1669,7 +1675,7 @@ class TestFIRBSupervisoryLGD:
                 "approach": [ApproachType.FIRB.value],
             }
         )
-        result = _pad(lf).irb.apply_firb_lgd(basel31_config).collect()
+        result = _pad(lf).pipe(apply_firb_lgd, basel31_config).collect()
         assert result["lgd"][0] == pytest.approx(0.40)
 
     def test_namespace_firb_lgd_airb_retains_own_lgd(
@@ -1685,7 +1691,7 @@ class TestFIRBSupervisoryLGD:
                 "approach": [ApproachType.AIRB.value],
             }
         )
-        result = _pad(lf).irb.apply_firb_lgd(basel31_config).collect()
+        result = _pad(lf).pipe(apply_firb_lgd, basel31_config).collect()
         assert result["lgd"][0] == pytest.approx(0.30)
 
     # --- FSE vs non-FSE LGD (Art. 161(1)(a) vs (aa)) ---
@@ -1704,7 +1710,7 @@ class TestFIRBSupervisoryLGD:
                 "cp_is_financial_sector_entity": [True],
             }
         )
-        result = _pad(lf).irb.apply_firb_lgd(basel31_config).collect()
+        result = _pad(lf).pipe(apply_firb_lgd, basel31_config).collect()
         assert result["lgd"][0] == pytest.approx(0.45)
 
     def test_namespace_b31_non_fse_firb_lgd_40pct(
@@ -1721,7 +1727,7 @@ class TestFIRBSupervisoryLGD:
                 "cp_is_financial_sector_entity": [False],
             }
         )
-        result = _pad(lf).irb.apply_firb_lgd(basel31_config).collect()
+        result = _pad(lf).pipe(apply_firb_lgd, basel31_config).collect()
         assert result["lgd"][0] == pytest.approx(0.40)
 
     def test_namespace_crr_fse_ignored(
@@ -1738,7 +1744,7 @@ class TestFIRBSupervisoryLGD:
                 "cp_is_financial_sector_entity": [True],
             }
         )
-        result = _pad(lf).irb.apply_firb_lgd(crr_config).collect()
+        result = _pad(lf).pipe(apply_firb_lgd, crr_config).collect()
         assert result["lgd"][0] == pytest.approx(0.45)
 
     # --- Covered bond LGD ---
